Update blueprint for deploying DICE-Traffic
The blueprint deploys a minimal Kafka - Cassandra
cluster complete with the needed Kafka connectors to automatically
write the results to Cassandra keyspaces.

Consider the blueprint an experiment about what minimal
changes and updates are needed to create a non-trivial
Big Data cluster and application.

For example, we needed to add relationships to connect a node
template to Kafka or Cassandra, obtaining the target's FQDN in
the runtime parameter. With Cloudify, this is not possible to
perform through runtime properties alone.

Change-Id: I3516ad8eb7a41c4596678a792a2b70fd4457f6f6