Fix the paths
Creation of the path for local data was incomplete, because
the tool expects an existing subfolder `/data`.
The data path needed to have a trailing delimeter `/` because
a certain tool concatenates the strings without trying to
add delimiters itself.
Also the path to the `pytraffic` tool was incorrect in the
systemd `.service` units.

Change-Id: I671e6bfaf72d2b9d7118039f910a05717ac1a217