Add recipes for creating VPN tunnel to Arctur
TIMON bits run in Arctur, so the new recipes `vpn_install` and
`vpn_start`, respectively, install + configure the OpenVPN client,
and start the client. We baked in the public parts of the needed
certificates. The private key does not belong to git, so the
user has to supply it as an attribute.

Change-Id: Id641aec95c7f5bbe9684a912aba8ae5d6a6fd0ba