Add Jenkinsfile for Jenkins pipeline builds
Being code-by-design extends to unit tests. Jenkins accepts
Jenkinsfiles where the definition of the job is baked into
the repository.
The new Jenkinsfile has two major stages. First it runs
unit tests of the Python project. Second, it executes
kitchen tests to validate the Chef cookbooks. This is
possible because for each stage we can specify the label
of the Jenkins node to take over the test.

Change-Id: I47f2c3a9faf9d7f421db7323e41538c9432dfbe7